Use ByteBuf instead of ByteBuffer

+++ legacy/src/main/kotlin/rs/emulate/legacy/widget/script/LegacyInstructionType.kt
@@ -147,8 +147,8 @@ enum class LegacyInstructionType(
     /**
      * Gets the integer value of this LegacyInstructionType.
      */
-    fun toInteger(): Short {
-        return value.toShort()
+    fun toInteger(): Int {
+        return value
     }
 
     companion object {
